reorganization of angles_* modules
